requirements.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
joblib==1.3.2
pyyaml==6.0.1
streamlit==1.28.1
plotly==5.17.0
requests==2.31.0